Fix compile errors in the server and client.
The client needs to define the type of struct since instances were already allocated in the definition of the grpc_metadata array. Also, the default nullptr's can be removed from the nested obfuscated struct's creation in the same statement.

The server needs to undef the Free macro defined by Rcpp so it can be defined by absl low_level_alloc later.